+++ NEWS.md
@@ -1,3 +1,24 @@
+# Kanta Harmonisation v1.2.0
+
+- Added new mappings from FinOMOP for the missing code with no unis
+
+#Kanta Harmonisation v1.1.0
+
+Major changes :
+
+- Added 2816 new mappings to LABfi_ALL.usagi.csv for lab codes with no unit, these were created from the existing mappings if all lab codes with different units mapped to same conceptId
+- modified check_lab_usagi_file, to create automatically mappings to lab codes with no units 
+
+# Kanta Harmonisation v1.0.0
+
+Major changes :
+
+- Update to work with summary of the real Finngen counts
+
+Minor changes:
+
+- Sam fixes in the usagi file 
+
 # Kanta Harmonisation v0.3.0
 
 - Updates by Sam, mostly 'ERROR; Units: Units dont match quantity'